Replace alert congratulations with sweetalert2
Had a hard time finding out how to integrate sweetalert2 to the code,
since just using it as it was mentioned in the usage required an import
that the current page doesn't allow since this project it's not module
based.

But using the "old-school approach" of initializing the plugin by
referencing the necessary files did the trick.
Of course for the script src itself I needed the link to the cdn of he
exact file.

Signed-off-by: Carolina Romero Carrasco <mcarolinarc28@gmail.com>